+++ sdk/core/Azure.Core/src/ClientOptions.cs
@@ -4,7 +4,9 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.Diagnostics.CodeAnalysis;
 using Azure.Core.Pipeline;
+using Microsoft.Extensions.Configuration;
 
 namespace Azure.Core
 {
@@ -63,8 +65,28 @@ namespace Azure.Core
                 // null as the argument rather than calling their default constructors. Calling their default constructors would result
                 // in a stack overflow as this constructor is called from a static initializer.
                 _transport = HttpPipelineTransport.Create();
-                Diagnostics = new DiagnosticsOptions(null);
-                Retry = new RetryOptions(null);
+                Diagnostics = new DiagnosticsOptions((DiagnosticsOptions?)null);
+                Retry = new RetryOptions((RetryOptions?)null);
+            }
+        }
+
+        /// <summary>
+        /// Creates a new instance of <see cref="ClientOptions"/> with the specified <see cref="IConfigurationSection"/>.
+        /// </summary>
+        /// <param name="section">The <see cref="IConfigurationSection"/> to read from.</param>
+        [Experimental("SCME0002")]
+        protected ClientOptions(IConfigurationSection section)
+        {
+            _transport = HttpPipelineTransport.Create();
+            if (section is null || !section.Exists())
+            {
+                Diagnostics = new DiagnosticsOptions((DiagnosticsOptions?)null);
+                Retry = new RetryOptions((RetryOptions?)null);
+            }
+            else
+            {
+                Diagnostics = new DiagnosticsOptions(section.GetSection("Diagnostics"));
+                Retry = new RetryOptions(section.GetSection("Retry"));
             }
         }