+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/CHANGELOG.md
@@ -48,7 +48,7 @@
 
 ## 1.6.0-beta.1 (2025-12-03)
 
-### Bugs Fixed
+### Features Added
 * Added Microsoft override attributes to preserve exact Application Insights
   semantics when exporting telemetry data.
   ([#54023](https://github.com/Azure/azure-sdk-for-net/pull/54023))